refactor: ♻️ merge and fix
