+++ src/lib/components/list/ProducaoBibliograficaList.svelte
@@ -1,0 +1,3 @@
+<script lang="ts">
+    
+</script>
