--- src/lib/components/wrapper.svelte
@@ -1,2 +1,0 @@
-
-			<slot/>
