+++ src/routes/+page.ts
@@ -5,8 +5,8 @@ export const load: PageLoad = async ({ url: { searchParams } }) => {
 	const campus = searchParams.get('campus');
 	const grandeArea = searchParams.get('grande_area');
 	const area = searchParams.get('area');
-	const anoLte = searchParams.get('ano_lte') ?? new Date().getFullYear().toString();
-	const anoGte = searchParams.get('ano_gte') ?? '2000'
+	// const anoLte = searchParams.get('ano_lte') ?? new Date().getFullYear().toString();
+	// const anoGte = searchParams.get('ano_gte') ?? '2000'
 
 	return {
 		grandesAreas: ProducoesService.grandeAreas({
@@ -17,15 +17,15 @@ export const load: PageLoad = async ({ url: { searchParams } }) => {
 			grandeArea
 		}),
 		campus: ProducoesService.campus(),
-		producoes: ProducoesService.list({
-			campus,
-			grandeArea,
-			area,
-			anoLte: anoLte ? parseInt(anoLte) : undefined,
-			anoGte: anoGte ? parseInt(anoGte) : undefined,
-			page: 1,
-			perPage: 10
-		}),
+		// producoes: ProducoesService.list({
+		// 	campus,
+		// 	grandeArea,
+		// 	area,
+		// 	anoLte: anoLte ? parseInt(anoLte) : undefined,
+		// 	anoGte: anoGte ? parseInt(anoGte) : undefined,
+		// 	page: 1,
+		// 	perPage: 10
+		// }),
 		chart: await ProducoesService.chart({
 			campus,
 			grandeArea,
